[configuration] extract loading into own module
