+++ apicast/libexec/boot
@@ -3,7 +3,7 @@
 pcall(require, 'luarocks.loader')
 package.path = package.path .. ";./src/?.lua"
 
-local configuration = require 'configuration'
+local configuration = require 'configuration_loader'
 
 local config = configuration.boot()
