+++ spec/configuration_spec.lua
@@ -21,33 +21,6 @@ describe('Configuration object', function()
     end)
   end)
 
-  describe('.download', function()
-    it('returns error on missing endpoint', function()
-      assert.same({nil, 'missing endpoint'}, { configuration.download() })
-    end)
-
-    it('returns error on invalid URI', function()
-      assert.same({nil, 'invalid endpoint'}, { configuration.download('foobar') })
-    end)
-
-    it('returns error on invalid URI', function()
-      assert.same({nil, 'connection refused'}, { configuration.download('http://127.0.0.1:1234/config/') })
-    end)
-  end)
-
-  it('.download', function()
-    configuration.download('http://user:pass@localhost:3000')
-    configuration.download('https://user@localhost')
-    configuration.download('http://:pass@lvh.me:3000')
-  end)
-
-  describe('.read', function()
-    it('ignores empty path', function()
-      assert.same({nil, 'missing path'}, { configuration.read() })
-      assert.same({nil, 'missing path'}, { configuration.read('') })
-    end)
-  end)
-
   describe('.parse_service', function()
     it('ignores empty hostname_rewrite', function()
       local config = configuration.parse_service({ proxy = { hostname_rewrite = '' }})
