+++ t/004-apicast-path-routing.t
@@ -24,7 +24,7 @@ env APICAST_PATH_ROUTING_ENABLED=1;
   include $TEST_NGINX_UPSTREAM_CONFIG;
   lua_package_path "$TEST_NGINX_LUA_PATH";
   init_by_lua_block {
-    require('configuration').save({
+    require('configuration_loader').save({
       services = {
         {
           id = 42,
