+++ t/005-apicast-oauth.t
@@ -24,7 +24,7 @@ __DATA__
   lua_package_path "$TEST_NGINX_LUA_PATH";
 
   init_by_lua_block {
-    require('configuration').save({
+    require('configuration_loader').save({
       services = {
         { backend_version = 'oauth',
           proxy = { oauth_login_url = "http://example.com/redirect" } }
@@ -49,7 +49,7 @@ Location: http://example.com/redirect?error=invalid_client
   lua_package_path "$TEST_NGINX_LUA_PATH";
 
   init_by_lua_block {
-    require('configuration').save({
+    require('configuration_loader').save({
       services = {
         { backend_version = 'oauth',
           proxy = { oauth_login_url = "http://example.com/redirect" } }
@@ -91,7 +91,7 @@ Location: http://example.com/redirect\?scope=whatever&response_type=code&state=\
   lua_package_path "$TEST_NGINX_LUA_PATH";
 
   init_by_lua_block {
-    require('configuration').save({
+    require('configuration_loader').save({
       services = {
         {
          id = 42, backend_version = 'oauth',
@@ -129,7 +129,7 @@ Location: http://example.com/redirect\?scope=whatever&response_type=token&error=
 --- http_config
   lua_package_path "$TEST_NGINX_LUA_PATH";
   init_by_lua_block {
-    require('configuration').save({
+    require('configuration_loader').save({
       services = {
         { backend_version = 'oauth' }
       }
@@ -147,7 +147,7 @@ POST /oauth/token
 --- http_config
   lua_package_path "$TEST_NGINX_LUA_PATH";
   init_by_lua_block {
-    require('configuration').save({
+    require('configuration_loader').save({
       services = {
         { backend_version = 'oauth' }
       }
@@ -165,7 +165,7 @@ POST /oauth/token?grant_type=authorization_code&client_id=client_id&redirect_uri
 --- http_config
   lua_package_path "$TEST_NGINX_LUA_PATH";
   init_by_lua_block {
-    require('configuration').save({
+    require('configuration_loader').save({
       services = {
         { backend_version = 'oauth' }
       }
@@ -183,7 +183,7 @@ GET /callback
 --- http_config
   lua_package_path "$TEST_NGINX_LUA_PATH";
   init_by_lua_block {
-    require('configuration').save({
+    require('configuration_loader').save({
       services = {
         { backend_version = 'oauth' }
       }
@@ -206,7 +206,7 @@ include $TEST_NGINX_APICAST_CONFIG;
   resolver $TEST_NGINX_RESOLVER;
   lua_package_path "$TEST_NGINX_LUA_PATH";
   init_by_lua_block {
-    require('configuration').save({
+    require('configuration_loader').save({
       services = {
         { backend_version = 'oauth' }
       }
@@ -228,7 +228,7 @@ Not part of the RFC. This is the Gateway API to create access tokens and redirec
   resolver $TEST_NGINX_RESOLVER;
   lua_package_path "$TEST_NGINX_LUA_PATH";
   init_by_lua_block {
-    require('configuration').save({
+    require('configuration_loader').save({
       services = {
         { id = 42, backend_version = 'oauth', oauth_login_url = "" }
       }
@@ -265,7 +265,7 @@ Location: http://example.com/redirect\?code=\w+&state=\w+
   resolver $TEST_NGINX_RESOLVER;
   lua_package_path "$TEST_NGINX_LUA_PATH";
   init_by_lua_block {
-    require('configuration').save({
+    require('configuration_loader').save({
       services = {
         { id = 42, backend_version = 'oauth' }
       }
@@ -332,7 +332,7 @@ GET /t
   include $TEST_NGINX_UPSTREAM_CONFIG;
 
   init_by_lua_block {
-    require('configuration').save({
+    require('configuration_loader').save({
       services = {
         {
           backend_version = 'oauth',
