+++ t/008-apicast-request-logs.t
@@ -25,7 +25,7 @@ __DATA__
   lua_package_path "$TEST_NGINX_LUA_PATH";
 
   init_by_lua_block {
-    require('configuration').save({
+    require('configuration_loader').save({
       services = {
         {
           id = 42,
@@ -77,7 +77,7 @@ env APICAST_REQUEST_LOGS=1;
   lua_package_path "$TEST_NGINX_LUA_PATH";
 
   init_by_lua_block {
-    require('configuration').save({
+    require('configuration_loader').save({
       services = {
         {
           id = 42,
@@ -134,7 +134,7 @@ env APICAST_RESPONSE_CODES=1;
   lua_package_path "$TEST_NGINX_LUA_PATH";
 
   init_by_lua_block {
-    require('configuration').save({
+    require('configuration_loader').save({
       services = {
         {
           id = 42,
@@ -192,7 +192,7 @@ env APICAST_RESPONSE_CODES=1;
   lua_package_path "$TEST_NGINX_LUA_PATH";
 
   init_by_lua_block {
-    require('configuration').save({
+    require('configuration_loader').save({
       services = {
         {
           id = 42,
